factored metrics logic out of the allocator main file
